Thêm api thống kê, sửa lại chức năng xóa thẻ và xóa vé, đã test toàn bộ chức năng trên postman

+++ bikerental/src/app/api/auth/card/deleteCard/route.js
@@ -2,11 +2,26 @@ import pool from "@/db.js";
 
 export async function DELETE(req) {
     try {
-        const { cardId } = await req.json();
+        const { the_id } = await req.json();
 
+        // Kiểm tra nếu the_id tồn tại trong cơ sở dữ liệu
+        const [rows] = await pool.execute(
+            "SELECT COUNT(*) AS count FROM `the` WHERE the_id = ?",
+            [the_id]
+        );
+
+        if (rows[0].count === 0) {
+            // Nếu không tồn tại, trả về thông báo thẻ không tồn tại
+            return new Response(
+                JSON.stringify({ message: "Thẻ không tồn tại!" }),
+                { status: 404 }
+            );
+        }
+
+        // Nếu tồn tại, tiến hành xóa thẻ
         await pool.execute(
             "DELETE FROM `the` WHERE the_id = ?",
-            [cardId]
+            [the_id]
         );
 
         return new Response(
