+++ bikerental/src/app/api/auth/tickets/deleteTickets/route.js
@@ -2,11 +2,25 @@ import pool from "@/db.js";
 
 export async function DELETE(req) {
     try {
-        const { ticketId } = await req.json();
+        const { ve_id } = await req.json();
+
+         // Kiểm tra nếu the_id tồn tại trong cơ sở dữ liệu
+         const [rows] = await pool.execute(
+            "SELECT COUNT(*) AS count FROM `ve` WHERE ve_id = ?",
+            [ve_id]
+        );
+
+        if (rows[0].count === 0) {
+            // Nếu không tồn tại, trả về thông báo thẻ không tồn tại
+            return new Response(
+                JSON.stringify({ message: "Vé không tồn tại!" }),
+                { status: 404 }
+            );
+        }
 
         await pool.execute(
             "DELETE FROM `ve` WHERE ve_id = ?",
-            [ticketId]
+            [ve_id]
         );
 
         return new Response(
